requirements.txt
pyspark==2.4.0
kafka-python